% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gitGrep}
\alias{gitGrep}
\title{A wrapper around the the \code{git grep -n <pattern>} command with a
find-and-replace option.}
\usage{
gitGrep(pattern, replacement = NA, path = ".", extra = "-nF")
}
\arguments{
\item{pattern}{regular expression to search for.  See \code{glob2rx} to convert
a glob to regular expression}

\item{replacement}{character, if not NA then replace occurance of pattern with this
string}

\item{path}{the path to the git snapshot}

\item{extra,}{character, extra modifiers for \code{git grep} that must include at least'-n'
The default is "-nF" which means the pattern fixed (not regex), specify \code{extra}
without "F" if you want to pass a regular expression.}
}
\value{
a data frame as returned by \code{\link{read.gitGrep}}
}
\description{
A wrapper around the the \code{git grep -n <pattern>} command with a
find-and-replace option.
}

