% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.gitGrep}
\alias{read.gitGrep}
\title{Reads the output from git grep into a data frame.}
\usage{
read.gitGrep(data, filename = NA, noText = FALSE)
}
\arguments{
\item{data}{character vector of the result so \code{$git grep ...}}

\item{filename}{the fully qualifed filename to read, provide this if you have
saved the data in a text file}

\item{noText}{if TRUE then do not include original text in returned data frame}
}
\value{
a data frame of the match data [,c("file", "line", "text")]
}
\description{
Output generated by git grep as shown below
$git grep -n "regex pattern" > result.txt ror you can use the function
provided here \code{gitGrep}
}

